+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Additions :tada:
+
+- Added `BoundingRegionBuilder::toGlobeRectangle`.
+
 ### v0.34.0 - 2024-04-01
 
 ##### Breaking Changes :mega: